refactor: exclude saved locations from searching results

+++ src/main/java/icekubit/dto/LocationDto.java
@@ -2,11 +2,13 @@ package icekubit.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class LocationDto {
     private String name;
     private String country;
-    private double lat;
-    private double lon;
+    private BigDecimal lat;
+    private BigDecimal lon;
 
 }
