+++ src/main/java/icekubit/entity/Location.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "Locations")
 @Data
@@ -17,8 +19,8 @@ public class Location {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
     private String name;
-    private double latitude;
-    private double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     @ManyToOne
     @JoinColumn(name = "UserId", referencedColumnName = "ID")
     private User user;
